Rendering read-only DOI fields as hyperlinks
These point to dx.doi.org, and open in a new frame or window.
