+++ controllers/study.py
@@ -10,6 +10,7 @@ import time
 import nexson
 import tempfile
 import gzip
+from link import doi2url
 
 from externalproc import get_external_proc_dir_for_upload, invoc_status, \
     ExternalProcStatus, get_logger, get_conf, do_ext_proc_launch
@@ -368,6 +369,10 @@ def view():
     t = db.study
     rec = t(request.args(0)) or redirect(URL("create"))
     readonly = not auth.has_membership(role="contributor")
+    t.doi.label = "DOI"
+    # make read-only DOIs into proper hyperlinks
+    if readonly:
+        t.doi.represent = lambda v: A(doi2url(v), _href=doi2url(v), _target='_blank')
     ## t.focal_clade.readable = t.focal_clade.writable = False
     t.focal_clade_ottol.label = 'Focal clade'
     t.focal_clade_ottol.widget = SQLFORM.widgets.autocomplete(
@@ -791,17 +796,6 @@ def ref_from_doi():
     wrap this call to an external service here so that the jQuery won't be concerned about
     cross-site scripting in the AJAX code for updating the create form.
     """
-    def normalize_doi_for_url(raw):
-        lowercase = raw.lower()
-        if lowercase.startswith('doi:'):
-            raw = raw[4:]
-        elif lowercase.startswith('doi'):
-            raw = raw[3:]
-        elif lowercase.startswith('http://dx.doi.org/'):
-            raw = raw[18:]
-        if lowercase.endswith('.json'):
-            raw = raw[:-5]
-        return raw
     def format_citation(d):
         '''
         Parses the output of the procite (vnd.citationstyles.csl+json) format
@@ -857,7 +851,7 @@ def ref_from_doi():
         return
     raw = '/'.join(list(request.args))
     DOMAIN = 'http://dx.doi.org'
-    doi = normalize_doi_for_url(raw)
+    doi = links.normalize_doi_for_url(raw)
     #sys.stderr.write('About look up reference for the doi "%s"\n' % doi)
     RETURNS_OBJECT = True
     SUBMIT_URI = DOMAIN + '/' + doi
